requirements.txt
google-generativeai==0.7.2
python-dotenv==1.0.0
streamlit==1.31.1
requests==2.31.0
pytest==7.4.3
